+++ lib/features/folder_detail/presentation/folder_detail_page.dart
@@ -53,10 +53,14 @@ class _FolderDetailPageState
   void initState() {
     super.initState();
     apiToken = widget.argument.apiToken;
+    xApplication = widget.argument.xApplication;
+    savePhotoAfterShot = widget.argument.savePhotoAfterShot;
     environment = widget.argument.environtment ?? Evn.production;
     locale = widget.argument.locale;
     enableVersion2 = widget.argument.enableVersion2 ?? true;
+    isAICycle = widget.argument.isAICycle ?? true;
     controller.argument = widget.argument;
+
     callEngineSub = controller.damageResponseStream.stream.listen((p0) {
       if (p0 != null) {
         widget.onCallEngineSuccessfully?.call(p0);